Fetch discount by coupon

+++ src/main/java/com/sorted/portal/response/beans/FetchCartV2.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Builder
 public record FetchCartV2(
         @JsonProperty("total_count")
-        Integer totalCount,
+        Long totalCount,
         @JsonProperty("total_amount")
         BigDecimal totalAmount,
         @JsonProperty("free_delivery_diff")
